Use float threshold for dither

+++ src/ImageSharp/Dithering/Ordered/IOrderedDither.cs
@@ -21,7 +21,7 @@ namespace SixLabors.ImageSharp.Dithering
         /// <param name="x">The column index.</param>
         /// <param name="y">The row index.</param>
         /// <typeparam name="TPixel">The pixel format.</typeparam>
-        void Dither<TPixel>(ImageFrame<TPixel> image, TPixel source, TPixel upper, TPixel lower, byte threshold, int x, int y)
+        void Dither<TPixel>(ImageFrame<TPixel> image, TPixel source, TPixel upper, TPixel lower, float threshold, int x, int y)
             where TPixel : struct, IPixel<TPixel>;
     }
 }
